# excourt_re/excourt_re-master/system/app/static/swagger.yaml
openapi: 3.0.0
info:
  title: ExCourt API
  description: API for court management system
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /offercourt/upload:
    post:
      summary: Create a court offer
      tags:
        - Court Offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - my_id
                - court_id
              properties:
                my_id:
                  type: integer
                  description: ID of the user offering the court
                court_id:
                  type: integer
                  description: ID of the court being offered
                upload_time:
                  type: string
                  format: date-time
                  description: Offer time (defaults to current time)
      responses:
        '201':
          description: Court offer successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Court offer published successfully
                  data:
                    type: object
                    properties:
                      Offer_upload_id:
                        type: integer
        '400':
          description: Missing required parameters
        '402':
          description: User has already uploaded a request for this court
        '500':
          description: Server error
  /offercourt/record:
    post:
      summary: Accept a court offer
      tags:
        - Court Offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - my_id
                - court_id
              properties:
                my_id:
                  type: integer
                  description: ID of the user accepting the offer
                court_id:
                  type: integer
                  description: ID of the court being accepted
                upload_time:
                  type: string
                  format: date-time
                  description: Accept time (defaults to current time)
      responses:
        '201':
          description: Court offer successfully accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Court offer accepted successfully
                  data:
                    type: object
                    properties:
                      Accept_record_id:
                        type: integer
        '400':
          description: Missing required parameters
        '404':
          description: No pending offer found for this court
        '409':
          description: A similar court offer record already exists
        '500':
          description: Server error
  /offercourt/accept:
    post:
      summary: Confirm court offer acceptance
      tags:
        - Court Offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - my_id
                - court_id
                - recorder_id
              properties:
                my_id:
                  type: integer
                  description: ID of the user who offered the court
                court_id:
                  type: integer
                  description: ID of the court
                recorder_id:
                  type: integer
                  description: ID of the user accepting the offer
      responses:
        '200':
          description: Court offer successfully confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Court offer accepted successfully
        '400':
          description: Missing required parameters
        '404':
          description: No matching record found
        '500':
          description: Server error
  /offercourt/decline:
    post:
      summary: Decline a court offer
      tags:
        - Court Offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - my_id
                - court_id
                - recorder_id
              properties:
                my_id:
                  type: integer
                  description: ID of the user who offered the court
                court_id:
                  type: integer
                  description: ID of the court
                recorder_id:
                  type: integer
                  description: ID of the user accepting the offer
      responses:
        '200':
          description: Court offer successfully declined
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 送场请求已拒绝
        '400':
          description: Missing required parameters
        '404':
          description: No matching record found
        '500':
          description: Server error
  /offercourt/get_offer_records:
    post:
      summary: Get all court offer records for a user
      tags:
        - Court Offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - my_id
              properties:
                my_id:
                  type: integer
                  description: ID of the user
      responses:
        '200':
          description: Records successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 获取记录成功
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        Offer_record_id:
                          type: integer
                        Offer_responser_id:
                          type: integer
                        Offer_state:
                          type: string
                        Offer_uploader_court_id:
                          type: integer
                        Student_name:
                          type: string
                        Student_profileurl:
                          type: string
                        Student_nickname:
                          type: string
                        Student_credit:
                          type: integer
                        Student_level:
                          type: integer
                        Student_status:
                          type: string
        '400':
          description: Missing required parameters
        '500':
          description: Server error
  /offercourt/get_uploader_id:
    post:
      summary: Get the ID of the user who offered a court
      tags:
        - Court Offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - court_id
              properties:
                court_id:
                  type: integer
                  description: ID of the court
      responses:
        '200':
          description: Uploader ID successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      uploader_id:
                        type: integer
        '400':
          description: Missing court_id
        '404':
          description: No pending offer found for this court
        '500':
          description: Server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] 